ai-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sentence-transformers==3.3.1
numpy==2.1.3
psycopg2-binary==2.9.10
pgvector==0.3.6
python-dotenv==1.0.1
pydantic==2.10.3
torch==2.6.0
